backend/requirements.txt
# Core Flask dependencies
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2

# Database
psycopg2-binary==2.9.7
redis==4.6.0

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.8.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# AI Integration
openai==0.27.9

# Utilities
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Development
gunicorn==21.2.0 